scripts/requirements.txt
psycopg2-binary==2.9.9
supabase==2.10.0
python-dotenv==1.0.1
